+++ test/geos_ffi/conversion_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "minitest/autorun"
+require_relative "../../lib/rgeo-geojson"
+require_relative "../common/conversion_test"
+
+class GeosFFIConversionTest < Minitest::Test # :nodoc:
+  include RGeo::GeoJSON::Tests::Common::ConversionTests
+
+  def setup
+    @factory = RGeo::Geos.factory(native_interface: :ffi)
+  end
+end